# contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Travel Tour Booking API
  version: 0.16.0
  description: |
    API cho website bán tour du lịch Việt Nam, hai thị trường (DK, VN),
    hai ngôn ngữ (da, vi).
servers:
  - url: /
    description: Cùng gốc với ứng dụng
...
paths:
...
  /api/v1/{market}/products:
    get:
      tags: [products]
      operationId: listProducts
      summary: Danh sách sản phẩm của một thị trường
      description: |
        Chỉ trả sản phẩm **đã xuất bản ở thị trường này** và **đã dịch sang locale
        đang xem**. Sản phẩm thiếu bản dịch biến mất hoàn toàn — không listing,
        không tìm kiếm, không hiện bản `da` thay thế (`docs/02` mục 4).

        `totalItems` đếm trong phạm vi `(market, locale)`, nên **hai locale ra hai
        con số khác nhau**. Đúng như thiết kế: frontend hiển thị đúng số này chứ
        không hardcode.
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: region
          in: query
          required: false
          description: Slug miền **trong locale đang xem** — `nordvietnam` với `da`, `mien-bac` với `vi`.
          schema:
            type: string
        - name: destination
          in: query
          required: false
          description: |
            Slug điểm đến **trong locale đang xem**. Lọc theo điểm đến **chính**
            của sản phẩm, không lọc theo mọi điểm đến sản phẩm ghé qua — một
            tour xuyên Việt thuộc về một điểm đến khởi hành, không thuộc về tất cả.
          schema:
            type: string
        - name: theme
          in: query
          required: false
          description: |
            Slug chủ đề **trong locale đang xem**. **Lặp lại được** để lọc nhiều
            chủ đề: `?theme=trekking&theme=du-thuyen-song`. Không dùng dấu phẩy —
            dấu phẩy vỡ khi giá trị có chứa dấu phẩy (`docs/13` mục 6).

            Nhiều giá trị nghĩa là **HOẶC**: sản phẩm mang ít nhất một trong số đó.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: productType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ProductType'
        - name: q
          in: query
          required: false
          description: |
            Tìm trong tiêu đề, **không phân biệt dấu**: gõ `hoi an` ra `Hội An`,
            gõ `halong` ra `Hạ Long`. Bỏ dấu ở tầng CSDL bằng `f_unaccent`.
          schema:
            type: string
            maxLength: 100
        - name: sort
          in: query
          required: false
          description: |
            Mặc định `title,asc`. `title` sắp theo collation của locale đang xem
            (`da-DK-x-icu` xếp æ ø å **sau** z), sắp ở tầng CSDL chứ không ở Java.
          schema:
            $ref: '#/components/schemas/ProductSort'
        - name: page
          in: query
          required: false
          description: Đánh số từ 0. Phân trang theo offset — `docs/13` mục 6.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 60
            default: 24
      responses:
        '200':
          description: Một trang sản phẩm
          headers:
            Content-Language:
              schema: { type: string }
            Vary:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/v1/{market}/redirects/{type}/{slug}:
    get:
      tags: [products]
      operationId: resolveSlug
      summary: Slug cũ trỏ tới slug nào bây giờ
      description: |
        Đổi slug làm mọi liên kết cũ chết — liên kết trong email đã gửi, trong bài
        viết của người khác, trong dấu trang của khách. Bảng `slug_history` giữ
        slug cũ (`docs/12` mục 6.1); endpoint này là đường đọc của nó.

        **Chỉ trả slug mới khi bản ghi đích thật sự xem được ở
        `(market, locale)` này.** Đây là điểm quan trọng nhất: một chuyển hướng
        tới trang `404` tệ hơn hẳn một trang `404` thẳng — khách bị đưa đi một
        vòng rồi vẫn không thấy gì, và công cụ tìm kiếm ghi nhận một chuỗi
        chuyển hướng dẫn tới ngõ cụt.

        Vì thế nó kiểm đúng những điều kiện mà đường đọc công khai kiểm: đã dịch
        sang locale này, đã gán và đang bán ở thị trường này, chưa xoá mềm. Thiếu
        bất kỳ điều nào thì trả `404`, và frontend hiện trang không tìm thấy.

        Slug hiện tại gửi vào đây cũng trả `404`: nó không phải slug cũ, và
        chuyển hướng một URL về chính nó là một vòng lặp.
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: type
          in: path
          required: true
          description: |
            Loại bản ghi. `slug_history` dùng chung cho nhiều loại, và cùng một
            slug cũ có thể tồn tại ở hai loại khác nhau — nên loại là bắt buộc,
            không đoán được.
          schema:
            type: string
            enum: [PRODUCT, DESTINATION]
        - name: slug
          in: path
          required: true
          description: Slug **cũ**, của đúng locale ở `Accept-Language`.
          schema:
            type: string
      responses:
        '200':
          description: Slug hiện tại của bản ghi đó
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlugRedirect'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/v1/{market}/products/{slug}/stops:
    get:
      tags: [products]
      operationId: getProductStops
      summary: Các chặng dừng của lộ trình, kèm ảnh và video từng chặng
      description: |
        Khối "bản đồ lộ trình" ở tab Tổng quan — `docs/05` mục 6.1 vẽ nó là
        `[bản đồ lộ trình]  Nơi đến | Số đêm | Điểm nhấn`.

        **Dựng từ lịch trình, không phải một danh sách nhập riêng.** Mỗi điểm đến
        xuất hiện **một lần**, với số đêm bằng số ngày lịch trình ngủ ở đó và danh
        sách những ngày ấy. Cho biên tập viên nhập lại danh sách chặng là mời hai
        nguồn sự thật về cùng một lộ trình, và chỗ chúng lệch nhau sẽ là chỗ khách
        đọc.

        Ảnh và video của mỗi chặng là **của điểm đến**, không của sản phẩm: cùng
        một Sa Pa xuất hiện trong nhiều tour, và chụp lại ảnh cho từng tour là
        nhân bản cả tệp lẫn chứng từ giấy phép.

        Endpoint riêng chứ không nhồi vào `/itinerary`: hai khối này ở hai tab
        khác nhau, và tab Tổng quan không nên tải mô tả của mười lăm ngày để vẽ
        một danh sách năm dòng. `404` với `COMBO` và `DAY_TOUR`, cùng lý do với
        `/itinerary`.
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Các chặng, theo thứ tự xuất hiện trong lịch trình
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProductStop' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/{market}/products/{slug}/itinerary:
    get:
      tags: [products]
      operationId: getItinerary
      summary: Lịch trình từng ngày
      description: |
        **Trả `404` với `COMBO` và `DAY_TOUR`** — hai loại đó không có lịch trình
        theo ngày (`docs/13` mục 9.1). Đây là 404 nói "tài nguyên này không tồn
        tại với loại sản phẩm này", không phải "sản phẩm không tồn tại".

        Ngày bay hoặc ngày trên tàu thì `destination` bị bỏ khỏi JSON.
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lịch trình, sắp theo số ngày tăng dần
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItineraryDay'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/{market}/products/{slug}/hotels:
    get:
      tags: [products]
      operationId: getHotelStays
      summary: Các chặng nghỉ khách sạn
      description: |
        **Trả `404` với `CRUISE`** — du thuyền ngủ trên tàu, và tab tương ứng
        đổi thành Tàu và cabin (`docs/05` mục 2).

        `name` của khách sạn **không dịch**: tên riêng giữ nguyên ở mọi ngôn ngữ
        (`docs/24` mục 5). Chỉ `description` mới phụ thuộc locale.
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Danh sách chặng nghỉ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelStay'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/{market}/products/{slug}/departures:
    get:
      tags: [products]
      operationId: listDepartures
      summary: Ngày khởi hành còn bán
      description: |
        **Không bao giờ được cache** (`docs/13` mục 8). Chỗ còn thay đổi từng
        phút; hiện số chỗ cũ là dẫn khách vào một giao dịch sẽ thất bại ở bước cuối.

        `status` là **giá trị tính ra**, không phải cột trong cơ sở dữ liệu:
        `GUARANTEED` suy từ số khách đã đặt so với ngưỡng đảm bảo khởi hành
        (`docs/14` mục 5). `seatsAvailable` đã **trừ chỗ đang giữ**.

        `CRUISE` trả **phẳng**, mỗi hạng cabin một dòng; gộp theo ngày là việc
        của frontend (`docs/13` mục 9.1).
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ngày khởi hành, sắp theo ngày tăng dần
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Departure'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/v1/{market}/products/{slug}:
    get:
      tags: [products]
      operationId: getProduct
      summary: Chi tiết một sản phẩm
      description: |
        `slug` **phụ thuộc locale**: cùng một tour có slug khác nhau ở `da` và
        `vi`, và slug của locale này không mở được ở locale kia.

        Trả về **đúng một** đối tượng con theo `productType` (ADR-005). Nhờ
        `discriminator`, generator sinh ra kiểu tổng — frontend quên xử lý một
        loại là lỗi biên dịch, không phải màn hình trắng.

        `404` gộp ba tình huống có chủ ý: không tồn tại, chưa dịch cho locale
        này, chưa gán vào thị trường này. Phân biệt ra ngoài là rò rỉ thông tin
        về sản phẩm chưa mở bán (`docs/13` mục 5.1).
      parameters:
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: slug
          in: path
          required: true
          description: Không dấu ở cả hai ngôn ngữ.
          schema:
            type: string
      responses:
        '200':
          description: Sản phẩm
          headers:
            Content-Language:
              schema: { type: string }
            Vary:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
components:
  parameters:
    Market:
      name: market
      in: path
      required: true
      description: |
        Mã thị trường, chữ thường trong đường dẫn. Quyết định khách **mua** gì:
        catalog, bảng giá, tiền tệ, cổng thanh toán, ràng buộc pháp lý.
        Không suy ra từ locale.
      schema:
        type: string
        enum: [dk, vn]
...
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: true
      description: |
        Locale, quyết định khách **đọc** bằng tiếng gì. Độc lập hoàn toàn với
        market: khách Việt sống ở Đan Mạch mua ở `dk` nhưng đọc `vi`.
      schema:
        type: string
        enum: [da, vi]
  responses:
    Error:
      description: Lỗi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SlugRedirect:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          description: |
            Slug hiện tại, **của locale đã yêu cầu**. Frontend ghép nó với đoạn
            đường dẫn của locale đó để dựng URL đích.
          example: vietnam-fra-nord-til-syd
...
    ItineraryDay:
      type: object
      required: [dayNumber, title, description]
      properties:
        dayNumber: { type: integer, format: int32, minimum: 1 }
        title: { type: string }
        description: { type: string }
        destination:
          allOf: [{ $ref: '#/components/schemas/Ref' }]
          description: Nơi ngủ đêm. Bỏ khỏi JSON với ngày bay hoặc ngày trên tàu.
        hotelName:
          type: string
          description: Tên riêng, **không dịch**.
    HotelStay:
      type: object
      required: [name, nights, destination]
      properties:
        name:
          type: string
          description: Tên riêng của khách sạn, **không dịch** — `docs/24` mục 5.
        stars: { type: integer, format: int32, minimum: 1, maximum: 5 }
        nights: { type: integer, format: int32, minimum: 1 }
        destination: { $ref: '#/components/schemas/Ref' }
        description:
          type: string
          description: Phụ thuộc locale. Bỏ khỏi JSON khi chưa dịch.
        image: { type: string }
    DepartureStatus:
      type: string
      description: |
        Giá trị **tính ra**, không phải cột trong cơ sở dữ liệu. Thứ tự xét ở
        `docs/14` mục 5; `GUARANTEED` xét **trước** `FEW_SEATS`.
      enum: [OPEN, FEW_SEATS, GUARANTEED, SOLD_OUT, PENDING]
    Departure:
      type: object
      required: [id, departDate, returnDate, days, status, seatsAvailable]
      properties:
        id: { type: string, format: uuid }
        departDate: { type: string, format: date }
        returnDate: { type: string, format: date }
        days: { type: integer, format: int32 }
        status: { $ref: '#/components/schemas/DepartureStatus' }
        seatsAvailable:
          type: integer
          format: int32
          minimum: 0
          description: Đã trừ chỗ đang giữ — `docs/14` mục 6.1.
        priceFrom:
          allOf: [{ $ref: '#/components/schemas/Money' }]
        cabinCategory:
          type: string
          description: Chỉ `CRUISE`.
        departureCity:
          type: string
          description: Chỉ những thị trường có nhiều điểm khởi hành.
...
    ProductType:
      type: string
      description: |
        Sáu loại sản phẩm, mỗi loại có bảng con riêng và ràng buộc nghiệp vụ
        riêng — `docs/04`. `PRIVATE_TOUR` **không đặt trực tiếp được**: CTA là
        "Yêu cầu báo giá".
      enum:
        - GROUP_TOUR
        - INDIVIDUAL_PACKAGE
        - PRIVATE_TOUR
        - CRUISE
        - COMBO
        - DAY_TOUR
    ProductSort:
      type: string
      description: Trường và chiều sắp xếp, phân tách bằng dấu phẩy — `docs/13` mục 6.
      enum:
        - 'title,asc'
        - 'title,desc'
        - 'priceFrom,asc'
        - 'priceFrom,desc'
        - 'durationDays,asc'
        - 'durationDays,desc'
      default: 'title,asc'
    Ref:
      type: object
      description: Tham chiếu tới một thực thể khác — chỉ slug và tên, đã dịch sang locale đang xem.
      required: [slug, name]
      properties:
        slug:
          type: string
          example: nordvietnam
        name:
          type: string
          example: Nordvietnam
    ProductSummary:
      type: object
      description: |
        Bản rút gọn dùng cho listing. **Không** chứa hành trình, khách sạn hay
        ngày khởi hành — những thứ đó chỉ có ở trang chi tiết, và ngày khởi hành
        thì không bao giờ được cache (`docs/13` mục 8).
      required: [slug, title, productType, shortDescription, heroImage, heroImageAlt, region, destination, isNew, reviewCount]
      properties:
        slug: { type: string, example: vietnam-fra-nord-til-syd }
        title: { type: string, example: Vietnam fra nord til syd }
        productType: { $ref: '#/components/schemas/ProductType' }
        shortDescription: { type: string }
        heroImage: { type: string }
        heroImageAlt:
          type: string
          description: Bắt buộc, và **phụ thuộc locale** — chữ alt cũng là nội dung phải dịch.
        durationDays:
          type: integer
          format: int32
          description: Bỏ hẳn khỏi JSON với `DAY_TOUR` — loại đó đo bằng giờ.
        region: { $ref: '#/components/schemas/Ref' }
        destination: { $ref: '#/components/schemas/Ref' }
        priceFrom:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: |
            Giá **thấp nhất** của thị trường này, không phải giá đã quy đổi tỷ
            giá — mỗi thị trường có giá riêng do người nhập (`CLAUDE.md` điều 4).
            Chưa có bảng giá thì **bỏ hẳn** trường này, frontend hiện "Liên hệ".
        isNew: { type: boolean }
        rating:
          type: number
          format: double
          minimum: 1
          maximum: 5
        reviewCount: { type: integer, format: int32, minimum: 0 }
    ProductPage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProductSummary' }
        page: { type: integer, format: int32, minimum: 0 }
        size: { type: integer, format: int32, minimum: 1 }
        totalItems:
          type: integer
          format: int64
          minimum: 0
          description: |
            Đếm trong phạm vi `(market, locale)` sau khi đã áp bộ lọc. Đây là con
            số frontend phải hiển thị — "Xem tất cả 92 tour" không được hardcode.
        totalPages: { type: integer, format: int32, minimum: 0 }
...
    GroupTourDetail:
      description: Đoàn ghép có trưởng đoàn. `guaranteedThreshold` là số khách tối thiểu để chuyến chắc chắn khởi hành.
      allOf:
        - $ref: '#/components/schemas/ProductDetailBase'
        - type: object
          required: [minPax, maxPax, guaranteedThreshold, tourLeaderLanguage, fitnessLevel]
          properties:
            minPax: { type: integer, format: int32 }
            maxPax: { type: integer, format: int32 }
            guaranteedThreshold: { type: integer, format: int32 }
            tourLeaderLanguage:
              type: string
              description: |
                Ngôn ngữ của trưởng đoàn. **Không phải** locale khách đang đọc —
                khách Việt ở Đan Mạch đọc `vi` nhưng đi đoàn có trưởng đoàn nói `da`.
            fitnessLevel: { type: integer, format: int32, minimum: 1, maximum: 4 }
    IndividualPackageDetail:
      description: Gói cá nhân, khách tự chọn ngày trong cửa sổ linh hoạt.
      allOf:
        - $ref: '#/components/schemas/ProductDetailBase'
        - type: object
          required: [minPartySize, flexibleDateWindowDays]
          properties:
            minPartySize: { type: integer, format: int32 }
            flexibleDateWindowDays: { type: integer, format: int32 }
    PrivateTourDetail:
      description: |
        Tour riêng — **không đặt trực tiếp được**. Bấm CTA sinh một yêu cầu báo
        giá; giá cuối do tư vấn viên dựng.
      allOf:
        - $ref: '#/components/schemas/ProductDetailBase'
        - type: object
          required: [leadTimeDays, quoteValidDays]
          properties:
            leadTimeDays:
              type: integer
              format: int32
              description: Ngày khởi hành sớm hơn số ngày này thì trả `LEAD_TIME_NOT_MET`.
            quoteValidDays: { type: integer, format: int32 }
    CruiseDetail:
      description: Du thuyền. Ngày khởi hành trả **phẳng**, gộp theo ngày là việc của frontend.
      allOf:
        - $ref: '#/components/schemas/ProductDetailBase'
        - type: object
          required: [shipName, portCount]
          properties:
            shipName: { type: string }
            portCount: { type: integer, format: int32 }
...
    ProductDetail:
      description: |
        Kiểu tổng theo `productType` — ADR-005. Generator sinh ra union
        TypeScript và interface Java, nên **quên xử lý một loại là lỗi biên dịch**.
      oneOf:
        - $ref: '#/components/schemas/GroupTourDetail'
        - $ref: '#/components/schemas/IndividualPackageDetail'
        - $ref: '#/components/schemas/PrivateTourDetail'
        - $ref: '#/components/schemas/CruiseDetail'
        - $ref: '#/components/schemas/ComboDetail'
        - $ref: '#/components/schemas/DayTourDetail'
      discriminator:
        propertyName: productType
        mapping:
          GROUP_TOUR: '#/components/schemas/GroupTourDetail'
          INDIVIDUAL_PACKAGE: '#/components/schemas/IndividualPackageDetail'
          PRIVATE_TOUR: '#/components/schemas/PrivateTourDetail'
          CRUISE: '#/components/schemas/CruiseDetail'
          COMBO: '#/components/schemas/ComboDetail'
          DAY_TOUR: '#/components/schemas/DayTourDetail'
    Money:
      type: object
      description: |
        `amount` là **chuỗi**, không phải số JSON — số dấu phẩy động của
        JavaScript làm hỏng tiền. API không định dạng: không có "24.990 kr."
        ở bất kỳ đâu trong phản hồi.
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d+)?$'
          example: '24990.00'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: DKK
...
    MediaItem:
      type: object
      description: |
        Một tấm ảnh hoặc một video, nhìn từ bề mặt khách.
        
        Khác `GalleryImage` ở chỗ nó mang được cả video. `GalleryImage` giữ nguyên
        cho bộ ảnh sản phẩm — bộ đó hôm nay chỉ có ảnh, và đổi kiểu của nó là đổi
        phá vỡ tương thích cho một thứ chưa cần.
      required: [kind, url, alt, width, height]
      properties:
        kind: { $ref: '#/components/schemas/MediaKind' }
        url: { type: string }
        alt:
          type: string
          description: |
            Theo locale ở `Accept-Language`. Tệp **không có `alt` ở locale đang
            đọc thì không xuất hiện** — luật không fallback, `CLAUDE.md` quy tắc 3.
        width: { type: integer, format: int32 }
        height: { type: integer, format: int32 }
        contentType: { type: string, description: Chỉ có ở `VIDEO`. }
        durationSeconds: { type: integer, format: int32, description: Chỉ có ở `VIDEO`. }
        posterUrl:
          type: string
          description: |
            Ảnh bìa của video. Luôn có với `VIDEO` — `ck_media_video` bắt buộc nó
            ở tầng CSDL.
    ProductStop:
      type: object
      description: |
        Một chặng dừng của lộ trình — khối "bản đồ lộ trình" ở tab Tổng quan
        (`docs/05` mục 6.1).

        **Dựng từ lịch trình, không nhập riêng.** Nhập hai lần là hai chỗ lệch
        nhau, và chỗ lệch sẽ là chỗ khách đọc.
      required: [destination, nights, dayNumbers, media]
      properties:
        destination: { $ref: '#/components/schemas/Ref' }
        nights:
          type: integer
          format: int32
          description: Số đêm ngủ ở chặng này — đếm số ngày lịch trình trỏ vào điểm đến đó.
        dayNumbers:
          type: array
          items: { type: integer, format: int32 }
          description: Những ngày ứng với chặng này, để khách đối chiếu với tab Lịch trình.
        media:
          type: array
          items: { $ref: '#/components/schemas/MediaItem' }
          description: |
            Ảnh và video của điểm đến, theo thứ tự biên tập viên đã sắp. **Có thể
            rỗng** — chặng chưa có ảnh vẫn phải hiện ra.
...
    ErrorResponse:
      type: object
      description: Mã lỗi kèm tham số. Frontend dịch — backend không trả câu tiếng người.
      required: [code]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        params:
          type: object
          additionalProperties: true
        fields:
          type: array
          description: |
            Chỉ có ở `VALIDATION_FAILED`, và chỉ khi lỗi định vị được tới từng
            trường (`docs/13` mục 5). Luật liên trường không nằm ở đây — chúng có
            mã lỗi riêng, ví dụ `DURATION_DAYS_RULE_VIOLATED`.

            Rỗng thì **bỏ hẳn khỏi JSON**, không trả `[]` — quy tắc trường rỗng
            của `docs/13` mục 4.
          items:
            $ref: '#/components/schemas/FieldError'
        traceId:
          type: string
...